refactor component menu links
